Smarty
Realized that I really wanted to just have Smarty already set up in this starter app… which means I can use it in the installer. Also realized that I'd like to compartmentalize the LTI scripts away from the rest of the app, to make it easier to push updates from apps back to the starter repo.

Right now, something's up in the installer (I'm pretty sure the notification messages are borked) and it's not getting to the OAuth negotiation.