Fixed
Okay, so… now everything is template. And more of the LTI is isolated in `/lti/`. And a really stupid typo in the installer is fixed, so the API tokens are now acquired.

The only lingering concern is that it allows someone to _not_ set this up on a secure server, which will cause Canvas to balk. Which is annoying.