Use API of current instance
This _should_ only matter if the user is working on the test or beta Canvas instances, but `CanvasAPIviaLTI` is now processing the passed Canvas API Domain to create the `CanvasPest $api`.

In theory, after the beta and test instances have been refreshed, API keys issued on the main instance get transferred over, so everyone should be happy and working. In theory.