More robust instance-awareness
I think [this issue](https://github.com/smtech/canvashack/issues/8) should now be resolved. Monitoring to make sure that the logic is working in real life. The real fix is dependent on upgrading `smtech/reflexive-canvas-lti` to 1.0.6.

Also added a setup script to facilitate configuration — there’s a problem where the cCanvasHack schema is not getting loaded, which I need to remind myself how to resolve elegantly. ;)